+++ rsass/tests/spec/css/plain/error/expression/parentheses.rs
@@ -8,7 +8,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // wrong error
 fn test() {
     assert_eq!(
         runner().err("@use \'plain\'"),