+++ CHANGELOG.md
@@ -29,7 +29,7 @@ project adheres to
   - Made all color channels f64 instead of Rational (PR #199).
 * Fixed a bug where `clamp(..)` was sometimes evaluated to a value
   even though units wasn't comparable.
-* Improved parse error handling (PR #201, Issue #141).
+* Improved parse error handling (Issue #141, PR #201, PR #205).
   Many parse errors now match the dart sass error message.
   Also allow "loud" comments in more places.
 * Updated sass-spec test suite to 2024-10-10.